# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Absher Travel & Tourism API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /offers:
    get:
      operationId: listOffers
      tags: [offers]
      summary: List all offers
      parameters:
        - name: featured
          in: query
          schema:
            type: boolean
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of offers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Offer"
    post:
      operationId: createOffer
      tags: [offers]
      summary: Create an offer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OfferInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Offer"
...
  /offers/{id}:
    get:
      operationId: getOffer
      tags: [offers]
      summary: Get an offer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Offer details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Offer"
    patch:
      operationId: updateOffer
      tags: [offers]
      summary: Update an offer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OfferUpdate"
      responses:
        "200":
          description: Updated offer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Offer"
    delete:
      operationId: deleteOffer
      tags: [offers]
      summary: Delete an offer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    Offer:
      type: object
      required: [id, titleAr, titleEn, descriptionAr, descriptionEn, imageUrl, isActive, sortOrder, featured, createdAt, updatedAt]
      properties:
        id: { type: integer }
        titleAr: { type: string }
        titleEn: { type: string }
        descriptionAr: { type: string }
        descriptionEn: { type: string }
        imageUrl: { type: string }
        discountLabel: { type: ["string", "null"] }
        linkUrl: { type: ["string", "null"] }
        startDate: { type: ["string", "null"] }
        endDate: { type: ["string", "null"] }
        isActive: { type: boolean }
        sortOrder: { type: integer }
        price: { type: ["number", "null"] }
        currency: { type: ["string", "null"] }
        duration: { type: ["string", "null"] }
        destination: { type: ["string", "null"] }
        featured: { type: boolean }
        createdAt: { type: string }
        updatedAt: { type: string }
    OfferInput:
      type: object
      required: [titleAr, titleEn, descriptionAr, descriptionEn, imageUrl]
      properties:
        titleAr: { type: string }
        titleEn: { type: string }
        descriptionAr: { type: string }
        descriptionEn: { type: string }
        imageUrl: { type: string }
        discountLabel: { type: string }
        linkUrl: { type: string }
        startDate: { type: string }
        endDate: { type: string }
        isActive: { type: boolean }
        sortOrder: { type: integer }
        price: { type: number }
        currency: { type: string }
        duration: { type: string }
        destination: { type: string }
        featured: { type: boolean }
    OfferUpdate:
      type: object
      properties:
        titleAr: { type: string }
        titleEn: { type: string }
        descriptionAr: { type: string }
        descriptionEn: { type: string }
        imageUrl: { type: string }
        discountLabel: { type: string }
        linkUrl: { type: string }
        startDate: { type: string }
        endDate: { type: string }
        isActive: { type: boolean }
        sortOrder: { type: integer }
        price: { type: number }
        currency: { type: string }
        duration: { type: string }
        destination: { type: string }
        featured: { type: boolean }
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT